Removed the fwd and common
