--- core/specfem/source/common/types.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-namespace specfem::sources {
-
-enum class source_type {
-  vector_source, ///< Vector source
-  tensor_source, ///< Tensor source
-};
-
-}
