+++ core/specfem/source/dim2/source.hpp
@@ -5,8 +5,7 @@
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
 #include "source_time_function/interface.hpp"
-#include "specfem/source/common/types.hpp"
-#include "specfem/source/fwd.hpp"
+#include "specfem/source.hpp"
 #include "specfem_mpi/interface.hpp"
 #include "specfem_setup.hpp"
 #include "utilities/interface.hpp"
