+++ core/specfem/source/dim2/tensor_source/moment_tensor_source.cpp
@@ -1,8 +1,8 @@
-#include "moment_tensor_source.hpp"
 #include "enumerations/interface.hpp"
 #include "globals.h"
 #include "kokkos_abstractions.h"
 #include "source_time_function/interface.hpp"
+#include "specfem/source.hpp"
 #include "specfem_setup.hpp"
 // #include "utilities.cpp"
 #include "yaml-cpp/yaml.h"
