+++ core/specfem/source/dim2/tensor_source/moment_tensor_source.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "../tensor_source.hpp"
 #include "constants.hpp"
 #include "enumerations/interface.hpp"
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
 #include "source_time_function/interface.hpp"
+#include "specfem/source.hpp"
 #include "specfem_mpi/interface.hpp"
 #include "specfem_setup.hpp"
 #include "utilities/interface.hpp"
