+++ core/specfem/source/dim2/vector_source/external.cpp
@@ -1,6 +1,6 @@
-#include "external.hpp"
 #include "enumerations/interface.hpp"
 #include "source_time_function/interface.hpp"
+#include "specfem/source.hpp"
 #include "specfem_setup.hpp"
 #include <cmath>
