+++ core/specfem/source/dim2/vector_source/force_source.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "../vector_source.hpp"
 #include "enumerations/specfem_enums.hpp"
 #include "kokkos_abstractions.h"
 #include "quadrature/interface.hpp"
 #include "source_time_function/interface.hpp"
+#include "specfem/source.hpp"
 #include "specfem_mpi/interface.hpp"
 #include "specfem_setup.hpp"
 #include "utilities/interface.hpp"
